renaming to test vectors files
