+++ balloon-hash/benches/lib.rs
@@ -6,7 +6,7 @@ use test::Bencher;
 use balloon_hash::{Algorithm, Balloon, Params};
 use digest::array::Array;
 
-#[path = "../data/balloon_m.rs"]
+#[path = "../data/balloon_m_test_vectors.rs"]
 mod test_vectors;
 use test_vectors::{BALLOON_M_TEST_VECTORS, BalloonMTestVector};
